Renamed one word components and all relative files by adding prefix OC, removed turned off rule
