+++ .eslintrc.cjs
@@ -6,6 +6,5 @@ module.exports = {
   extends: ["eslint:recommended", "plugin:vue/vue3-recommended", "prettier"],
   rules: {
     "vue/require-default-prop": ["off"],
-    "vue/multi-word-component-names": ["off"],
   },
 };
