+++ packages/@orchidui-vue/src/index.js
@@ -1,3 +1,3 @@
-export * from "./ui/Button/Button";
-export * from "./ui/Chip/Chip";
-export * from "./ui/Dropdown/Dropdown";
+export * from "./ui/Button/OCButton.js";
+export * from "./ui/Chip/OCChip.js";
+export * from "./ui/Dropdown/OCDropdown.js";
