--- packages/@orchidui-vue/src/ui/Button/Button.js
@@ -1,2 +1,0 @@
-import Button from "./Button.vue";
-export { Button };
