+++ packages/@orchidui-vue/src/ui/Button/OCButton.js
@@ -1,0 +1,3 @@
+import Button from "./OCButton.vue";
+
+export { Button };
