--- packages/@orchidui-vue/src/ui/Chip/Chip.js
@@ -1,3 +1,0 @@
-import Chip from "./Chip.vue";
-
-export { Chip };
