+++ packages/@orchidui-vue/src/ui/Chip/OCChip.js
@@ -1,0 +1,3 @@
+import Chip from "./OCChip.vue";
+
+export { Chip };
