--- packages/@orchidui-vue/src/ui/Dropdown/Dropdown.js
@@ -1,3 +1,0 @@
-import Dropdown from "./Dropdown.vue";
-
-export { Dropdown };
