+++ packages/@orchidui-vue/src/ui/Dropdown/OCDropdown.js
@@ -1,0 +1,3 @@
+import Dropdown from "./OCDropdown.vue";
+
+export { Dropdown };
