+++ packages/@orchidui-vue/src/ui/Theme/OCTheme.js
@@ -1,0 +1,3 @@
+import Theme from "./OCTheme.vue";
+
+export { Theme };
