--- packages/@orchidui-vue/src/ui/Theme/Theme.js
@@ -1,3 +1,0 @@
-import Theme from "./Theme.vue";
-
-export { Theme };
